third_party/WebKit/LayoutTests/shadow-dom/slots.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<slot id='slot_in_document'></slot>

<script>
test(() => {
  assert_equals(window.getComputedStyle(slot_in_document).getPropertyValue('display'),
                'contents');
}, 'Slots: Blink specific tests for slots.');
</script>
